Refactor/estimation contract (#556)
* checkpoint

* refactor(EntryPointSimulations): update aggregator info handling

* refactor(entry-point): update to use new v8 interfaces

* refactor(src/v08): consolidate interface versions

* docs(README): update documentation for Alto Contracts

* refactor(api): rename functions for clarity and consistency

* feat(interface): add context simulation methods

* refactor(interface): rename context parameters for clarity

* style(contracts): forgefmt

* refactor(simulations): replace BinarySearchArgs with PackedUserOperation

* refactor(EntryPointSimulations): extract target call data encoding

* refactor(simulations): simplify user operation handling

* style(EntryPointSimulations): format code for readability

* refactor(interface): remove redundant simulateHandleOp methods

* refactor(contracts): rename and update PimlicoSimulations ABI

* refactor(gasEstimator): improve state overrides handling

* fix(paymaster): handle case with no paymaster data

* refactor(gasEstimator): rename methods for consistency

* refactor(gasEstimator): replace ContractFunctionExecutionError with ContractFunctionRevertedError

* feat(gasEstimator): add binarySearch method helpers

* refactor(gasEstimator): simplify gas limit handling logic

* refactor(gasEstimations): improve variable naming and error handling

* refactor(gasEstimator): simplify gas estimation logic

* refactor(gasEstimator): update return type for gas estimation methods

* refactor(gasEstimations): rename and simplify simulation result types

* refactor(gasEstimation): simplify error handling in gas estimators

* feat(gasEstimator): add gas estimation simulation method

* feat(binary-search): enhance binary search result structure

* refactor(gasEstimations): simplify gas estimation logic

* refactor(gasEstimations): simplify error handling in gas estimation

* refactor(gasEstimator): simplify gas limit search methods

* refactor(gasEstimations): replace searchType with methodName

* feat(gasEstimator): add error decoding for contract function reverts

* style(gasEstimations): change error logging to warnings

* refactor(gasEstimator): rename stateOverrides to userStateOverrides

* refactor(validation): rename userOperation to userOp

* refactor(rpc): standardize user operation parameter names

* style(contracts): clean up comments and formatting

* refactor(rpc): standardize userOperation parameter naming

* fix build

* style(rpc): remove unnecessary parentheses in gas limit calculation

* test(eth_estimateUserOperationGas): add tests for gas limits

* refactor(validation): simplify state override handling

* refactor(userop): improve type safety in stateOverrides mapping

* refactor(gasEstimation): rename userOp to userOperation

* refactor: rename gas limit functions for clarity

* refactor(contracts): rename gas estimation functions and types

* refactor(gasEstimations): unify error handling for gas estimation

* refactor(estimation): standardize state override handling

* style(rpc): improve code readability in gas estimation

* feat(utils): return empty array for empty userOps

* refactor(gasEstimator): simplify error handling in gas estimation

* feat(simulations): add simulation functions for gas limits

* feat(simulation): add user operation simulation functions

* feat(gasEstimations): add new simulation contracts for gas estimation

* refactor(gasEstimator): simplify simulation contract handling

* refactor(rpc): update import path for parseFailedOpWithRevert

* refactor(validation): update simulation ABI references in SafeValidator

* refactor(simulations): update simulateEntryPoint to improve readability

* style(simulations): rename identifier to selector for clarity

* refactor(gasEstimations): update ABI and error handling

* refactor(console): remove unused console logs

* refactor(test): simplify user operation preparation in tests

* test(eth_estimateUserOperationGas): update test to use callData

* fix(simulations): improve error messages in revert handling

* style(BinarySearch): add comments to clarify unused results

* feat(simulation): add bulk entry point simulation function

* refactor(simulations): rename simulateEntryPoint to _simulateEntryPoint

* feat(gasEstimator): add state override preparation method

* refactor(estimation): consolidate state override preparation logic

* refactor(rpc): simplify error handling in gas estimators

* refactor(gasEstimations): simplify gas estimation logic

* refactor(contracts): rename EntryPointV06SimulationsAbi

* refactor(contracts): remove EntryPointSimulations06Abi

* fix(estimation): handle specific error case in decodeSimulateHandleOpError

* test(eth_estimateUserOperationGas): update error message in test

* refactor(gasEstimations): improve error logging in simulateValidation

---------

+++ contracts/src/misc/CodeHashGetter.sol
@@ -2,15 +2,15 @@
 pragma solidity ^0.8.15;
 
 contract CodeHashGetter {
-
     error CodeHashesResult(bytes32 hash);
+
     constructor(address[] memory addresses) {
         revert CodeHashesResult(getCodeHashes(addresses));
     }
 
     function getCodeHashes(address[] memory addresses) public view returns (bytes32) {
         bytes32[] memory hashes = new bytes32[](addresses.length);
-        for (uint i = 0; i < addresses.length; i++) {
+        for (uint256 i = 0; i < addresses.length; i++) {
             hashes[i] = addresses[i].codehash;
         }
         bytes memory data = abi.encode(hashes);
